restructure modules for packaging
