+++ tests/configs/test_service_config.py
@@ -4,11 +4,11 @@ from dataclasses import asdict
 from pathlib import Path
 
 import pytest
-from configs.service_config import load_service_config_from_file
-from exceptions import ConfigNotFoundError
-from exceptions import ConfigParseError
-from exceptions import ConfigValidationError
 
+from devservices.configs.service_config import load_service_config_from_file
+from devservices.exceptions import ConfigNotFoundError
+from devservices.exceptions import ConfigParseError
+from devservices.exceptions import ConfigValidationError
 from tests.testutils import create_config_file
 
 
@@ -64,7 +64,7 @@ def test_load_service_config_from_file(
     }
     create_config_file(tmp_path, config)
 
-    service_config = load_service_config_from_file(tmp_path)
+    service_config = load_service_config_from_file(str(tmp_path))
     assert asdict(service_config) == {
         "version": 0.1,
         "service_name": service_name,
@@ -86,7 +86,7 @@ def test_load_service_config_from_file_no_dependencies(tmp_path: Path) -> None:
     }
     create_config_file(tmp_path, config)
 
-    service_config = load_service_config_from_file(tmp_path)
+    service_config = load_service_config_from_file(str(tmp_path))
     assert asdict(service_config) == {
         "version": 0.1,
         "service_name": "example-service",
@@ -97,7 +97,7 @@ def test_load_service_config_from_file_no_dependencies(tmp_path: Path) -> None:
 
 def test_load_service_config_from_file_missing_config(tmp_path: Path) -> None:
     with pytest.raises(ConfigNotFoundError) as e:
-        load_service_config_from_file(tmp_path)
+        load_service_config_from_file(str(tmp_path))
     assert (
         str(e.value)
         == f"Config file not found in directory: {tmp_path / 'devservices' / 'docker-compose.yml'}"
@@ -118,7 +118,7 @@ def test_load_service_config_from_file_invalid_version(tmp_path: Path) -> None:
     create_config_file(tmp_path, config)
 
     with pytest.raises(ConfigValidationError) as e:
-        load_service_config_from_file(tmp_path)
+        load_service_config_from_file(str(tmp_path))
     assert str(e.value) == "Invalid version '0.2' in service config"
 
 
@@ -134,7 +134,7 @@ def test_load_service_config_from_file_missing_version(tmp_path: Path) -> None:
     create_config_file(tmp_path, config)
 
     with pytest.raises(ConfigValidationError) as e:
-        load_service_config_from_file(tmp_path)
+        load_service_config_from_file(str(tmp_path))
     assert str(e.value) == "Version is required in service config"
 
 
@@ -151,7 +151,7 @@ def test_load_service_config_from_file_missing_service_name(tmp_path: Path) -> N
     create_config_file(tmp_path, config)
 
     with pytest.raises(ConfigValidationError) as e:
-        load_service_config_from_file(tmp_path)
+        load_service_config_from_file(str(tmp_path))
     assert str(e.value) == "Service name is required in service config"
 
 
@@ -169,7 +169,7 @@ def test_load_service_config_from_file_invalid_dependency(tmp_path: Path) -> Non
     create_config_file(tmp_path, config)
 
     with pytest.raises(ConfigValidationError) as e:
-        load_service_config_from_file(tmp_path)
+        load_service_config_from_file(str(tmp_path))
     assert (
         str(e.value)
         == "Service 'unknown-dependency' in mode 'default' is not defined in dependencies"
@@ -190,7 +190,7 @@ def test_load_service_config_from_file_missing_default_mode(tmp_path: Path) -> N
     create_config_file(tmp_path, config)
 
     with pytest.raises(ConfigValidationError) as e:
-        load_service_config_from_file(tmp_path)
+        load_service_config_from_file(str(tmp_path))
     assert str(e.value) == "Default mode is required in service config"
 
 
@@ -207,7 +207,7 @@ def test_load_service_config_from_file_no_modes(tmp_path: Path) -> None:
     create_config_file(tmp_path, config)
 
     with pytest.raises(ConfigValidationError) as e:
-        load_service_config_from_file(tmp_path)
+        load_service_config_from_file(str(tmp_path))
     assert str(e.value) == "Default mode is required in service config"
 
 
@@ -228,7 +228,7 @@ def test_load_service_config_from_file_invalid_dependencies(tmp_path: Path) -> N
     create_config_file(tmp_path, config)
 
     with pytest.raises(ConfigParseError) as e:
-        load_service_config_from_file(tmp_path)
+        load_service_config_from_file(str(tmp_path))
     assert (
         str(e.value)
         == "Error parsing service dependencies: Dependency.__init__() got an unexpected keyword argument 'unknown'"
@@ -252,7 +252,7 @@ def test_load_service_config_from_file_invalid_modes(tmp_path: Path) -> None:
     create_config_file(tmp_path, config)
 
     with pytest.raises(ConfigValidationError) as e:
-        load_service_config_from_file(tmp_path)
+        load_service_config_from_file(str(tmp_path))
     assert str(e.value) == "Services in mode 'custom' must be a list"
 
 
@@ -272,7 +272,7 @@ def test_load_service_config_from_file_no_x_sentry_service_config(
     create_config_file(tmp_path, config)
 
     with pytest.raises(ConfigParseError) as e:
-        load_service_config_from_file(tmp_path)
+        load_service_config_from_file(str(tmp_path))
     assert str(e.value) == "Config file does not contain 'x-sentry-service-config' key"
 
 
@@ -292,7 +292,7 @@ def test_load_service_config_from_file_invalid_yaml(tmp_path: Path) -> None:
         f.write(config)
 
     with pytest.raises(ConfigParseError) as e:
-        load_service_config_from_file(tmp_path)
+        load_service_config_from_file(str(tmp_path))
     assert (
         str(e.value)
         == f"Error parsing config file: mapping values are not allowed here\n  in \"{tmp_path / 'devservices' / 'docker-compose.yml'}\", line 2, column 12"
@@ -316,7 +316,7 @@ def test_load_service_config_from_file_invalid_yaml_tag(tmp_path: Path) -> None:
         f.write(config)
 
     with pytest.raises(ConfigParseError) as e:
-        load_service_config_from_file(tmp_path)
+        load_service_config_from_file(str(tmp_path))
     assert (
         str(e.value)
         == f"Error parsing config file: could not determine a constructor for the tag 'tag:yaml.org,2002:invalid_tag'\n  in \"{tmp_path / 'devservices' / 'docker-compose.yml'}\", line 7, column 19"
